alter some unit tests to use randomly generated batches
