+++ packages/rollup-contracts/test/helpers/index.ts
@@ -1,11 +1,3 @@
-/* Imports */
-import {
-  keccak256,
-  abi,
-  hexStrToBuf,
-  bufToHexString,
-} from '@eth-optimism/core-utils'
-
 /**********************************
  * Byte String Generation Helpers *
  *********************************/
@@ -18,74 +10,14 @@ export function makeRepeatedBytes(value: string, length: number): string {
   return '0x' + sliced
 }
 
-// Make padded bytes. Bytes are right padded.
-export function makePaddedBytes(value: string, length: number): string {
-  if (value.length > length * 2) {
-    throw new Error('Value too large to fit in ' + length + ' byte string')
+export function makeRandomBlockOfSize(blockSize: number): string[] {
+  const block = []
+  for (let i = 0; i < blockSize; i++) {
+    block.push(makeRepeatedBytes('' + Math.floor(Math.random() * 500 + 1), 32))
   }
-  const targetLength = length * 2
-  while (value.length < (targetLength || 2)) {
-    value = value + '0'
-  }
-  return '0x' + value
+  return block
 }
 
-// Make a padded uint. Uints are left padded.
-export function makePaddedUint(value: string, length: number): string {
-  if (value.length > length * 2) {
-    throw new Error('Value too large to fit in ' + length + ' byte string')
-  }
-  const targetLength = length * 2
-  while (value.length < (targetLength || 2)) {
-    value = '0' + value
-  }
-  return '0x' + value
+export function makeRandomBatchOfSize(batchSize: number): string[] {
+  return makeRandomBlockOfSize(batchSize)
 }
-
-/*******************************
- * Transition Encoding Helpers *
- ******************************/
-// export type Transition = string
-
-// // Generates some number of dummy transitions
-// export function generateNTransitions(
-//   numTransitions: number
-// ): RollupTransition[] {
-//   const transitions = []
-//   for (let i = 0; i < numTransitions; i++) {
-//     const transfer: TransferTransition = {
-//       stateRoot: getStateRoot('ab'),
-//       senderSlotIndex: 2,
-//       recipientSlotIndex: 2,
-//       tokenType: 0,
-//       amount: 1,
-//       signature: getSignature('01'),
-//     }
-//     transitions.push(transfer)
-//   }
-//   return transitions
-// }
-
-/****************
- * Misc Helpers *
- ***************/
-
-export const ZERO_BYTES32 = makeRepeatedBytes('0', 32)
-export const ZERO_ADDRESS = makeRepeatedBytes('0', 20)
-export const ZERO_UINT32 = makeRepeatedBytes('0', 4)
-export const ZERO_SIGNATURE = makeRepeatedBytes('0', 65)
-
-/* Extra Helpers */
-export const STORAGE_TREE_HEIGHT = 5
-export const AMOUNT_BYTES = 5
-export const getSlot = (storageSlot: string) =>
-  makePaddedUint(storageSlot, STORAGE_TREE_HEIGHT)
-export const getAmount = (amount: string) =>
-  makePaddedUint(amount, AMOUNT_BYTES)
-export const getAddress = (address: string) => makeRepeatedBytes(address, 20)
-export const getSignature = (sig: string) => makeRepeatedBytes(sig, 65)
-export const getStateRoot = (bytes: string) => makeRepeatedBytes(bytes, 32)
-export const getBytes32 = (bytes: string) => makeRepeatedBytes(bytes, 32)
-
-export const UNISWAP_ADDRESS = getAddress('00')
-export const UNISWAP_STORAGE_SLOT = 0
